Move setup to utils
